feat: analysis
