+++ ui/src/pages/analysis/components/Results/index.module.scss
@@ -42,3 +42,7 @@
 .result-content {
   margin-top: 16px;
 }
+
+.no-data {
+  padding-bottom: 100px;
+}
