--- ui/src/pages/targeting/components/TargetingForm/index.module.scss
@@ -1,184 +1,0 @@
-.status {
-  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.01), 0 3px 6px 3px rgba(0, 0, 0, 0.01), 0 2px 6px 0 rgba(0, 0, 0, 0.03);
-  border-radius: 12px;
-  background-color: #ffffff;
-  padding: 24px;
-}
-
-.joyride-status {
-  display: inline-flex;
-  width: auto;
-}
-
-.toggle-status {
-  display: flex;
-  font-size: 13px;
-  color: #212529;
-  align-items: center;
-  margin-left: 12px;
-}
-
-.status-text {
-  line-height: 28px;
-  margin-left: 20px;
-  font-size: 13px;
-  color: #74788d;
-  height: 28px;
-}
-
-.name-color {
-  display: inline-block;
-  width: 12px;
-  height: 12px;
-  margin-left: 5px;
-  border-radius: 4px;
-  vertical-align: middle;
-  position: relative;
-  top: -1px;
-}
-
-.name-text {
-  vertical-align: middle;
-  margin-left: 4px;
-  display: inline-block;
-  max-width: 400px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  position: relative;
-  top: -2px;
-}
-
-.variations {
-  margin-top: 12px;
-  background-color: #ffffff;
-  border-radius: 12px;
-  padding: 12px 20px 12px 24px;
-}
-
-.toggle-status-text {
-  margin-right: 12px;
-}
-
-.rules {
-  margin-top: 12px;
-  background-color: #ffffff;
-  border-radius: 12px;
-  padding: 0 24px 12px;
-}
-
-.disabled-return {
-  border-radius: 12px 12px 0 0;
-  background-color: #ffffff;
-  margin-top: 12px;
-  padding: 0 24px 12px;
-}
-
-.content-item {
-  padding-top: 24px;
-}
-
-.footer {
-  height: 56px;
-  background: #ffffff;
-  border-top: 1px solid #ebedef;
-  line-height: 56px;
-  padding: 0 20px;
-  position: sticky;
-  bottom: 0;
-  right: 30px;
-  left: 220px;
-  z-index: 1;
-}
-
-.publish-btn {
-  margin-right: 8px !important;
-}
-
-.publish-btn-loader {
-  margin-right: 5px !important;
-}
-
-.publish-btn-text {
-  vertical-align: middle;
-}
-
-.modal-header {
-  display: flex;
-  justify-content: space-between;
-  font-family: PingFangSC-Medium;
-  font-size: 16px;
-  color: #212529;
-  margin-bottom: 16px;
-  line-height: 26px;
-}
-
-.modal-close-icon {
-  cursor: pointer;
-}
-
-.modal-header-text {
-  font-family: PingFangSC-Medium;
-  font-size: 16px;
-  color: rgba(0, 0, 0, 0.87);
-}
-
-.modal-content {
-  font-size: 14px;
-  color: rgba(0, 0, 0, 0.60);
-  margin: 0 -16px -4px;
-  border-bottom: 1px solid rgba(206, 212, 218, 0.6);
-  padding-bottom: 24px;
-}
-
-.comment {
-  padding: 0 24px;
-  align-items: center;
-}
-
-.comment-title {
-  margin-bottom: 8px;
-}
-
-.comment-input {
-  :global {
-    .ui.input {
-      width: 100%;
-    }
-
-    .ui.input > input {
-      height: 32px;
-    }
-  }
-}
-
-.approval {
-  padding: 0 24px;
-  margin-bottom: 24px;
-}
-
-.approval-content {
-  display: flex;
-  align-items: center;
-}
-
-.approval-dropdown {
-  flex: 1;
-}
-
-.approval-btn {
-  margin-left: 8px;
-  color: #556ee6;
-  cursor: pointer;
-}
-
-.icon-info {
-  font-size: 14px !important;
-  color: #74788d;
-  vertical-align: middle;
-  margin-left: 6px;
-}
-
-.diff-after {
-  margin-top: 12px;
-}
