+++ ui/src/pages/targeting/index.module.scss
@@ -1,52 +1,136 @@
-.targeting {
-  position: relative;
-  padding: 0 20px;
-  min-width: 1200px;
+.status {
+  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.01), 0 3px 6px 3px rgba(0, 0, 0, 0.01), 0 2px 6px 0 rgba(0, 0, 0, 0.03);
+  border-radius: 12px;
+  background-color: #ffffff;
+  padding: 24px;
 }
 
-.menus {
+.joyride-status {
+  display: inline-flex;
+  width: auto;
+}
+
+.toggle-status {
   display: flex;
+  font-size: 13px;
+  color: #212529;
   align-items: center;
-}
-
-.menu {
-  margin: 12px 0 !important;
-  flex: 1;
-}
-
-.put-away {
-  vertical-align: bottom;
-  margin-right: 4px;
-}
-
-.content {
-  display: flex;
-  position: relative;
-  min-height: calc(100vh - 366px);
-}
-
-.content-left {
-  flex: 1;
-}
-
-.content-right {
-  overflow-y: auto;
-  box-shadow: 0 10px 20px 0 rgba(18, 38, 63, 0.03);
-  background: #ffffff;
-  border-radius: 8px 8px 0 0;
-  width: 260px;
   margin-left: 12px;
 }
 
-.popup {
-  padding: 0 !important;
-  margin-top: 4px !important;
+.status-text {
+  line-height: 28px;
+  margin-left: 20px;
+  font-size: 13px;
+  color: #74788d;
+  height: 28px;
 }
 
-.warning-circle {
-  font-size: 20px !important;
-  color: #ffc300;
-  padding: 0 10px;
+.name-color {
+  display: inline-block;
+  width: 12px;
+  height: 12px;
+  margin-left: 5px;
+  border-radius: 4px;
+  vertical-align: middle;
+  position: relative;
+  top: -1px;
+}
+
+.name-text {
+  vertical-align: middle;
+  margin-left: 4px;
+  display: inline-block;
+  max-width: 400px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  position: relative;
+  top: -2px;
+}
+
+.variations {
+  margin-top: 12px;
+  background-color: #ffffff;
+  border-radius: 12px;
+  padding: 12px 20px 12px 24px;
+}
+
+.toggle-status-text {
+  margin-right: 12px;
+}
+
+.rules {
+  margin-top: 12px;
+  background-color: #ffffff;
+  border-radius: 12px;
+  padding: 0 24px 12px;
+}
+
+.disabled-return {
+  border-radius: 12px 12px 0 0;
+  background-color: #ffffff;
+  margin-top: 12px;
+  padding: 0 24px 12px;
+}
+
+.content-item {
+  padding-top: 24px;
+}
+
+.footer {
+  height: 56px;
+  background: #ffffff;
+  border-top: 1px solid #ebedef;
+  line-height: 56px;
+  padding: 0 20px;
+  position: sticky;
+  bottom: 0;
+  right: 30px;
+  left: 220px;
+  z-index: 1;
+}
+
+.publish-btn {
+  margin-right: 8px !important;
+}
+
+.publish-btn-loader {
+  margin-right: 5px !important;
+}
+
+.publish-btn-text {
+  vertical-align: middle;
+}
+
+.modal-header {
+  display: flex;
+  justify-content: space-between;
+  font-family: PingFangSC-Medium;
+  font-size: 16px;
+  color: #212529;
+  margin-bottom: 16px;
+  line-height: 26px;
+}
+
+.modal-close-icon {
+  cursor: pointer;
+}
+
+.publish-tips {
+  background: #fff5f4;
+  border-radius: 6px;
+  font-size: 12px;
+  color: #660e16;
+  margin: 0 24px 24px;
+  line-height: 32px;
+  display: flex;
+}
+
+.error-circle {
+  font-size: 17px !important;
+  color: #f5483b;
+  margin: 0 8px 0 14px;
 }
 
 .modal-header-text {
@@ -57,80 +141,78 @@
 
 .modal-content {
   font-size: 14px;
-  color: rgba(0, 0, 0, 0.6);
-  margin-top: 8px;
-  margin-left: 40px;
-  width: 320px;
+  color: rgba(0, 0, 0, 0.60);
+  margin: 0 -16px -4px;
+  border-bottom: 1px solid rgba(206, 212, 218, 0.6);
+  padding-bottom: 24px;
 }
 
-.message-content {
-  padding: 8px 16px;
-  border-radius: 6px;
-  font-size: 14px;
-  margin-bottom: 12px;
-  height: 36px;
+.collect {
+  padding: 0 24px;
+  align-items: center;
+}
+
+.collect-title {
+  margin-bottom: 8px;
+}
+
+.collect-content {
+  display: flex;
+  margin-left: 12px;
+}
+
+.collect-radio {
+  margin-right: 50px !important;
+}
+
+.comment {
+  padding: 0 24px;
+  align-items: center;
+}
+
+.comment-title {
+  margin-bottom: 8px;
+}
+
+.comment-input {
+  :global {
+    .ui.input {
+      width: 100%;
+    }
+
+    .ui.input > input {
+      height: 32px;
+    }
+  }
+}
+
+.approval {
+  padding: 0 24px;
+  margin-bottom: 24px;
+}
+
+.approval-content {
+  display: flex;
+  align-items: center;
+}
+
+.approval-dropdown {
   flex: 1;
-  position: relative;
 }
 
-.icon-warning-circle {
-  color: #ffc300;
-  vertical-align: middle;
-  font-size: 20px;
-}
-
-.message-content-warn {
-  background: #fffae0;
-  color: #592d00;
-}
-
-.message-icon {
-  margin-right: 8px;
-  font-size: 20px;
+.approval-btn {
   margin-left: 8px;
-}
-
-.message-content-text {
-  font-size: 14px;
-  margin-left: 8px;
-}
-
-.close-icon {
-  position: absolute;
-  right: 10px;
-  font-size: 18px;
+  color: #556ee6;
   cursor: pointer;
 }
 
-.archived-img {
-  position: fixed;
-  right: -0;
-  top: 310px;
-  width: 160px;
-  z-index: 1;
+.icon-info {
+  font-size: 14px !important;
+  color: #74788d;
+  vertical-align: middle;
+  margin-left: 6px;
 }
 
-.permanent-message-box {
-  height: 0;
-  display: flex;
-  justify-content: center;
-}
-
-.permanent-warning-message {
-  display: inline-flex;
-  align-items: center;
-  position: relative;
-  top: -42px;
-  background-color: #fffae0;
-  border-radius: 8px;
-  height: 33px;
-  font-size: 13px;
-  padding: 0 12px;
-  z-index: 9999;
-}
-
-.warning-icon {
-  display: inline-block;
-  color: #ffc300;
-  margin-right: 8px;
+.diff-after {
+  margin-top: 12px;
 }
