+++ ui/src/pages/toggleDetail/components/FlowExplain/index.module.scss
