+++ ui/src/pages/toggleDetail/components/Info/index.module.scss
@@ -1,0 +1,322 @@
+  
+.info {
+  background-color: #ffffff;
+  padding: 16px 24px;
+  box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.01), 0 3px 6px 3px rgba(0, 0, 0, 0.01), 0 2px 6px 0 rgba(0, 0, 0, 0.03);
+  border-radius: 0 0 12px 12px;
+  position: relative;
+  min-height: 188px;
+}
+  
+.info-title {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 28px;
+  margin-bottom: 24px;
+  line-height: 28px;
+}
+
+.info-title-left {
+  flex: 1;
+  display: flex;
+  align-items: center;
+}
+
+.toggle-lock-bg {
+  width: 24px;
+  height: 24px;
+  display: inline-block;
+  border-radius: 100%;
+  text-align: center;
+  line-height: 24px;
+  margin-right: 8px;
+}
+
+.toggle-lock-bg:hover {
+  background: rgba(33, 37, 41, 0.04);
+}
+
+.toggle-lock {
+  color: #f5483b;
+  font-size: 16px !important;
+}
+
+.popup-line {
+  margin-bottom: 8px;
+  font-size: 13px;
+  color: #eef2f7;
+}
+
+.popup-line:last-child {
+  margin-bottom: 0;
+}
+
+.status {
+  display: inline-block;
+  height: 26px;
+  border-radius: 16px;
+  line-height: 24px;
+  font-family: PingFangSC-Medium;
+  font-size: 12px;
+  padding: 0 8px;
+  margin-left: 8px;
+}
+
+.status-pending {
+  background: #eff8ff;
+  color: #4070f4;
+}
+
+.status-publish {
+  background: #e5fbf0;
+  color: #00b365;
+}
+
+.status-reject {
+  background: #feecef;
+  color: #ff4462;
+}
+
+.status-icon {
+  position: relative;
+  top: 1px;
+  margin-right: 4px;
+}
+
+.info-title-right {
+  display: flex;
+  align-items: center;
+}
+
+.btn {
+  margin-left: 8px !important;
+  min-width: 88px;
+}
+
+.dangerous-btn {
+  margin-left: 8px !important;
+  color: #f5483b !important;
+  border: 1px solid #f5483b !important;
+  min-width: 88px;
+}
+
+.dangerous-btn:hover {
+  background-color: transparent !important;
+}
+
+.loader {
+  margin-right: 5px !important;
+}
+
+.loader-text {
+  vertical-align: middle;
+}
+
+.connect-sdk {
+  color: #556ee6;
+  cursor: pointer;
+  margin-right: 8px;
+}
+
+.icon-connect-sdk {
+  margin-right: 6px !important;
+}
+
+.info-toggle-name {
+  font-family: HelveticaNeue-Medium;
+  font-size: 20px;
+  color: #212529;
+}
+
+.info-toggle-status {
+  margin-left: 8px;
+}
+
+.info-toggle-tags {
+  margin-left: 16px;
+}
+
+.tags {
+  background: #ececf6;
+  border-radius: 4px;
+  font-size: 10px;
+  color: #495057;
+  display: inline-block;
+  margin: 4px 8px 4px 0;
+  padding: 4px;
+}
+
+.tags-text {
+  transform: scale(0.8);
+  line-height: 1;
+}
+
+.info-content-row {
+  margin-top: 12px;
+}
+
+.label {
+  color: #74788d;
+  font-size: 13px;
+}
+
+.label-value {
+  font-size: 13px;
+  color: #495057;
+  font-family: HelveticaNeue-Medium;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.label-value-copy {
+  cursor: pointer;
+}
+
+.modal-header {
+  display: flex;
+  justify-content: space-between;
+  font-family: PingFangSC-Medium;
+  font-size: 16px;
+  color: #212529;
+}
+
+.modal-header-icon {
+  font-size: 16px !important;
+  cursor: pointer;
+  color: #495057;
+}
+
+.modal-content {
+  margin-top: 15px;
+}
+
+.modal-continue-edit {
+  font-family: PingFangSC-Regular;
+  font-size: 13px;
+  color: #495057;
+}
+
+.radio-group {
+  display: flex;
+  margin-top: 8px;
+
+  :global {
+    label {
+      color: #212529 !important;
+    }
+  }
+}
+
+.radio-group-item {
+  margin-right: 80px !important;
+}
+
+.input {
+  margin-top: 8px !important;
+}
+
+.label-required {
+  color: #ff5f57;
+  font-size: 20px;
+  margin-right: 4px;
+  height: 16px;
+  display: inline-block;
+  position: relative;
+  top: 5px;
+}
+
+.error-text {
+  color: #f46a6a;
+  margin-top: -15px;
+  margin-bottom: 15px;
+  font-size: 12px;
+  transform: scale(0.83);
+  transform-origin: left;
+}
+
+.footer {
+  margin-top: 20px;
+  text-align: right;
+}
+
+.footer-btn {
+  margin-left: 8px !important;
+}
+
+.diff-modal-header {
+  display: flex;
+  justify-content: space-between;
+  font-family: PingFangSC-Medium;
+  font-size: 16px;
+  color: #212529;
+  margin-bottom: 16px;
+  line-height: 26px;
+}
+
+.diff-modal-close-icon {
+  cursor: pointer;
+}
+
+.diff-modal-header-text {
+  font-family: PingFangSC-Medium;
+  font-size: 16px;
+  color: rgba(0,0,0,0.87);
+}
+
+.diff-modal-content {
+  font-size: 14px;
+  color: rgba(0, 0, 0, 0.60);
+  margin: 0 -16px -4px;
+
+  :global {
+    .d2h-file-header {
+      display: none;
+    }
+
+    .d2h-info {
+      display: none;
+    }
+
+    .d2h-file-wrapper {
+      border-top: 1px solid rgba(206, 212, 218, 0.6);
+      border-bottom: 1px solid rgba(206, 212, 218, 0.6);
+      border-radius: 0;
+    }
+
+    .d2h-del {
+      background-color: #fff5f4;
+    }
+
+    .d2h-code-line del, .d2h-code-side-line del {
+      background-color: #ffd4d0;
+    }
+
+    .d2h-ins {
+      background-color: #effcf2;
+    }
+
+    .d2h-code-line ins, .d2h-code-side-line ins {
+      background-color: #c2f6ce;
+    }
+
+    .d2h-file-side-diff {
+      margin-bottom: -2px;
+      overflow-x: auto;
+    }
+
+    .d2h-files-diff {
+      max-height: 341px;
+      overflow-y: auto;
+      z-index: 1;
+      position: relative;
+    }
+  }
+}
+
+.warinig-circle {
+  font-size: 20px !important;
+  color: #ff5f57;
+  margin-right: 8px;
+}
