+++ ui/src/semantic-ui/site/modules/popup.overrides
@@ -15,3 +15,11 @@
 .ui.popup:before {
   box-shadow: none;
 }
+
+.ui.bottom.left.popup:before {
+  box-shadow: none;
+}
+
+.ui.bottom.right.popup:before {
+  box-shadow: none;
+}
